Add support for common yang-library code between restconf and netconf
The yang-library support code has been moved from the apteryx-rest repo
to the apteryx-xml repo, so it can be used as shared code with netconf.

Model information for the hello message is now derived directly from
the yang-library data in apteryx.

Removed the -s option for a supported model list.

The help message capabilities are now derived from the yang-library
data.