Improve the handling of "with-defaults" options.
This change modifies the way "with-defaults" "report-all" is handled.
A new routine sch_add_defaults in apteryx-xml/schema.c is now called
to handle the complexity of adding defaults.

The routine _sch_xml_to_gnode was modified so the returned rschema node
matches that found by similar xpath and restconf code.