Fix with-defaults processing
Two tests were added to replicate recent issues discovered. Other tests
have been altered to pass as a result of the changes in schemas for the
test suites having been changed.

In the routine get_response_node, it is possible to change the variable
rschema (response schema), but the corresponding depth of this node
should also be changed at this time.